Switch to fakes3; get s3 working in docker
This implements a HostPortS3ConnectionContext which we need for doing a
local fake-s3 kind of thing. I'm tempted to merge this into the
S3ConnectionContext and have it do "modes" depending on whether
host/port are set, but I'm not sure if that overcomplicates things or
not. I'm not thrilled with the class hierarchy here.

This replaces the minio docker container with a fakes3 one. The minio
one has a bunch of issues and I couldn't get it to work in the context
of the version of boto we're using, docker and socorro-collector code.
The fakes3 one is literally a fake s3, so it ignores all kinds of stuff.
Further, the project seems kind of dead--there's a bunch of issues and
prs that are just sitting around. I think it's ok to use for now, but we
might want to switch to something else in the future. it's probably ok
to use, though.